+++ i18n.js
@@ -24,11 +24,11 @@ module.exports = {
     '/interactive-exercise/[slug]': ['exercises', 'workshops'],
     '/choose-program': ['choose-program', 'dashboard', 'profile', 'assignments'],
     '/syllabus/[cohortSlug]/[lesson]/[lessonSlug]': ['syllabus', 'dashboard', 'projects', 'assignments'],
-    '/subscription/[subId]': ['subscription'],
     '/survey/[surveyId]': ['survey'],
     '/mentorship': ['mentorship'],
     '/mentorship/schedule': ['mentorship', 'dashboard', 'signup', 'common'],
     '/how-to': ['how-to'],
+    '/paymentmethod': ['payment'],
     '/pricing': ['pricing', 'signup'],
     '/how-to/[slug]': ['how-to'],
     '/profile': ['profile'],
